+++ CHANGELOG.md
@@ -1,7 +1,11 @@
-## 0.3.1-dev2
+## 0.3.1-dev4
 
 ### Enhancements
+
 * **Add `precheck` to Milvus connector**
+* **LanceDB V2 Destination Connector**
+* **Persist record id in milvus, use it to delete previous content to prevent duplicates.**
+* **Persist record id in weaviate metadata, use it to delete previous content to prevent duplicates.**
 
 ### Fixes
 * **Make AstraDB precheck fail on non-existant collections**
@@ -27,7 +31,6 @@
 ## 0.2.2
 
 ### Enhancements
-
 * **Remove `overwrite` field** from fsspec and databricks connectors
 * **Added migration for GitLab Source V2**
 * **Added V2 confluence source connector**
@@ -44,6 +47,7 @@
 * **Astra DB V2 Source Connector** Create a v2 version of the Astra DB Source Connector.
 * **Support native async requests from unstructured-client**
 * **Support filtering element types in partitioner step**
+
 
 ### Fixes